Add `file_locator` to allow tracking files other than by inode
`file_locator` allows us to refer to a file in the filesystem such that we
can later call getxattr on it if necessary.  It generalises over the
different forms the functions take (e.g. stat, fstat, lstat, fstatat) by
always converting to the most general `fstatat` form.

This will allow me in future commits to add a new fakeroot backend based
on xattrs rather than a central daemon database.