get_stat: Don't error if we don't have permissions to call getxattr
Sometimes tools create files with mode 0000 with the intention of `chmod()`
them to the real mode later.  This caused problems for us because
`getxattr()` would fail with `EACCES`.  We know that if we have this data
stored in the xattrs the file will not have mode 0000 because our `chown()`
always gives us read/write permission, so don't bother attempting to read
xattrs if we don't have permission.